mtd: spi-nor: Add a flash_info entry for Macronix mx25uw51245g
mx25uw51245g is a SPI NOR that supports the SFDPRD command but returns
an empty SFDP page. This forces us to add a new entry in the flash_info
table.

Note that the NOR supports modes 1-1-1 and 8-8-8, but nothing in
between, hence the absence of SPI_NOR_{DUAL,QUAD}_READ flags.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>